fix(frontend): extract WeatherWidget fetch to weatherService (#348)
WeatherWidget.svelte made a direct fetch() to the Open-Meteo API,
bypassing the centralized API client pattern. Extracted the fetch
logic, caching, and geolocation into a dedicated weatherService.ts
in lib/services/.

Changes:
- Created lib/services/weatherService.ts: encapsulates Open-Meteo API
  call with 15-min weather cache and 24-hr geolocation cache
- Refactored WeatherWidget.svelte: now imports fetchWeather() from
  weatherService, removing ~80 lines of inline fetch/cache logic
- Added browser guards to all localStorage access in weatherService

The component is now a pure UI consumer; all API/cache logic is in
the service layer, consistent with the rest of the codebase.

Closes #348
